refactor for maintainability
