+++ backend/danswer/indexing/models.py
@@ -108,6 +108,8 @@ class EmbeddingModelDetail(BaseModel):
         embedding_model: "EmbeddingModel",
     ) -> "EmbeddingModelDetail":
         return cls(
+            # When constructing EmbeddingModel Detail for user-facing flows, strip the
+            # unneeded additional data after the `_`s
             model_name=embedding_model.model_name.split("__")[0],
             model_dim=embedding_model.model_dim,
             normalize=embedding_model.normalize,
