+++ backend/danswer/server/manage/search_settings.py
@@ -55,14 +55,10 @@ def set_new_embedding_model(
 
         embed_model_details.cloud_provider_id = cloud_id
 
+    # account for same model name being indexed with two different configurations
     if embed_model_details.model_name == current_model.model_name:
         if "__" not in current_model.model_name:
             embed_model_details.model_name += "__alt"
-
-        # raise HTTPException(
-        #     status_code=status.HTTP_400_BAD_REQUEST,
-        #     detail="New embedding model is the same as the currently active one.",
-        # )
 
     secondary_model = get_secondary_db_embedding_model(db_session)
