+++ web/src/app/admin/embeddings/components/EmbeddingSidebar.tsx
