+++ web/src/app/admin/embeddings/components/ModelSelector.tsx
