+++ web/src/app/admin/embeddings/components/ReindexingProgressTable.tsx
