+++ web/src/app/admin/embeddings/components/types.ts
