+++ web/src/app/admin/embeddings/modals/AlreadyPickedModal.tsx
