+++ web/src/app/admin/embeddings/modals/DeleteCredentialsModal.tsx
