+++ web/src/app/admin/embeddings/modals/ModelSelectionModal.tsx
