+++ web/src/app/admin/embeddings/modals/ProviderCreationModal.tsx
@@ -5,7 +5,7 @@ import * as Yup from "yup";
 import { Label, TextFormField } from "@/components/admin/connectors/Field";
 import { LoadingAnimation } from "@/components/Loading";
 import { CloudEmbeddingProvider } from "../components/types";
-import { EMBEDDING_PROVIDERS_ADMIN_URL } from "../../llm/constants";
+import { EMBEDDING_PROVIDERS_ADMIN_URL } from "../../configuration/llm/constants";
 import { Modal } from "@/components/Modal";
 
 export function ProviderCreationModal({
