+++ web/src/app/admin/embeddings/page.tsx
@@ -1,5 +1,18 @@
-import EmbeddingWrapper from "./EmbeddingWrapper";
+"use client";
 
-export default async function Page({}: {}) {
-  return <EmbeddingWrapper />;
+import { EmbeddingFormProvider } from "@/components/context/EmbeddingContext";
+import EmbeddingSidebar from "./components/EmbeddingSidebar";
+import EmbeddingForm from "./pages/EmbeddingForm";
+
+export default function EmbeddingWrapper() {
+  return (
+    <EmbeddingFormProvider>
+      <div className="flex justify-center w-full h-full">
+        <EmbeddingSidebar />
+        <div className="mt-12 w-full max-w-5xl mx-auto">
+          <EmbeddingForm />
+        </div>
+      </div>
+    </EmbeddingFormProvider>
+  );
 }
